+++ docs/HISTORY.rst
@@ -2,18 +2,20 @@
 History
 =======
 
-0.1.0 (2022-XX-XX)
+0.1.0 (2022-03-XX)
 ------------------
 * Build tests for all modules
 * Integrated slcpy, spindletorch and graphformer
     * SLCPY:
         * Loading of data types: *.tif, *.am, *.mrc, *.rec for 2D and 3D
-        * Include all slcpy modules
+        * Included all slcpy modules
         * Move Amira file output of point cloud from graphformer
         * SetUp workflows for data pre- and post-processing 
 
     * SPINDLETORCH
-        *
+        * Included all spindletorch modules
+        * Build standard workflows for training and prediction of 2D and 3D images
+
     * GRAPHFORMER
         *